+++ src/_includes/_includes/articleList.liquid
@@ -32,7 +32,9 @@
 
 {% if article.data.thumbnail %}
 <img src="{{article.data.thumbnail}}" class="db shadow-4" alt="{{article.data.title}}">
-{% elsif article.data.headerImage %}
+{% endif %}
+
+{% if article.data.headerImage %}
 {% image article.data.headerImage article.data.title 342 "db shadow-4" %}
 {% elsif article.data.gameImage %}
 <img src="{{article.data.gameImage}}" class="db shadow-4" alt="{{article.data.title}}">
